Cleaned up the structure of auth pages, removed unnecessary components. Fixed a bug where default mui style props for textFields included 'required' prop (should be optional).
